+++ src/main/java/shop/tryit/domain/answer/entity/Answer.java
@@ -1,4 +1,4 @@
-package shop.tryit.domain.answer;
+package shop.tryit.domain.answer.entity;
 
 import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.IDENTITY;
