+++ src/main/java/shop/tryit/domain/answer/repository/AnswerRepository.java
@@ -1,9 +1,10 @@
-package shop.tryit.domain.answer;
+package shop.tryit.domain.answer.repository;
 
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import shop.tryit.domain.answer.entity.Answer;
 import shop.tryit.domain.question.entity.Question;
 
 public interface AnswerRepository {
