+++ src/main/java/shop/tryit/domain/question/service/QnAFacade.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.domain.answer.Answer;
 import shop.tryit.domain.answer.dto.AnswerFormDto;
+import shop.tryit.domain.answer.entity.Answer;
 import shop.tryit.domain.answer.service.AnswerService;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberService;
