+++ src/main/java/shop/tryit/domain/question/service/QuestionService.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import shop.tryit.domain.answer.AnswerRepository;
+import shop.tryit.domain.answer.repository.AnswerRepository;
 import shop.tryit.domain.question.QuestionRepository;
 import shop.tryit.domain.question.dto.QuestionSearchCondition;
 import shop.tryit.domain.question.dto.QuestionSearchDto;
