+++ src/main/java/shop/tryit/repository/answer/AnswerJpaRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.tryit.domain.answer.Answer;
+import shop.tryit.domain.answer.entity.Answer;
 import shop.tryit.domain.question.entity.Question;
 
 public interface AnswerJpaRepository extends JpaRepository<Answer, Long> {
