+++ src/main/java/shop/tryit/repository/answer/AnswerRepositoryImpl.java
@@ -6,8 +6,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
-import shop.tryit.domain.answer.Answer;
-import shop.tryit.domain.answer.AnswerRepository;
+import shop.tryit.domain.answer.entity.Answer;
+import shop.tryit.domain.answer.repository.AnswerRepository;
 import shop.tryit.domain.question.entity.Question;
 
 @Repository
