+++ src/main/java/shop/tryit/repository/question/QuestionCustomImpl.java
@@ -1,7 +1,7 @@
 package shop.tryit.repository.question;
 
 import static java.util.stream.Collectors.toMap;
-import static shop.tryit.domain.answer.QAnswer.answer;
+import static shop.tryit.domain.answer.entity.QAnswer.answer;
 import static shop.tryit.domain.member.QMember.member;
 import static shop.tryit.domain.question.entity.QQuestion.question;
