+++ src/test/java/shop/tryit/repository/question/QuestionRepositoryImplTests.java
@@ -10,8 +10,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.domain.answer.Answer;
-import shop.tryit.domain.answer.AnswerRepository;
+import shop.tryit.domain.answer.entity.Answer;
+import shop.tryit.domain.answer.repository.AnswerRepository;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberRepository;
 import shop.tryit.domain.question.QuestionRepository;
